Ingestion: Refactor, docs, airflow example
